Crate iobc_supervisor_service

Kubos Service for interacting with the ISIS-OBC Supervisor

Configuration

The service must be configured in /etc/kubos-config.toml with the following fields:

[iobc-supervisor-service.addr]

ip - Specifies the service's IP address

port - Specifies the port on which the service will be listening for UDP packets

For example:

[iobc-supervisor-service.addr]
ip = "0.0.0.0"
port = 8170

Starting the Service

The service should be started automatically by its init script, but may also be started manually:

$ iobc-supervisor-service
Kubos antenna systems service started
Listening on: 0.0.0.0:8170

Available Fields

query {
    ping: "pong",
	supervisor: {
		version: {
			dummy,
			spiCommandStatus,
			indexOfSubsystem,
			majorVersion,
			minorVersion,
			patchVersion,
			gitHeadVersion,
			serialNumber,
			compileInformation,
			clockSpeed,
			codeType,
			crc
		},
		housekeeping: {
			dummy,
			spiCommandStatus,
			enableStatus: {
				powerObc,
				powerRtc,
				supervisorMode,
				busyRtc,
				powerOffRtc
			},
			supervisorUptime,
			iobcUptime,
			iobcResetCount,
			adcData,
			adcUpdateFlag,
			crc8
		}
	}
}

mutation {
	reset,
	emergencyReset,
	powercycle
}